requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-import-export==3.3.3
django-extensions==3.2.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
factory-boy==3.3.0
django-debug-toolbar==4.2.0
djangorestframework-api-key==3.0.0
django-ratelimit==4.1.0
channels==4.0.0
daphne==4.0.0
drf-yasg>=1.21.7,<2
python-dotenv==1.1.1
locust==2.34.0